Change evaluating implementation
Before this commit, we evaluate Node to Value by `eval` function.
This is very straightforward. But this approach does not work well
when implementing reference with store because we can make reference for
lambda value. When we deref location and get lambda value in the context of
"apply", we should apply the argument to the lambda. If we store the lambda
value as `Value`, we should de-construct the lambda value in `eval`
function. Introduce "value node" concept to make it easy to handle these situation.